refactor: Move CheckpointList into its own file
